feat: extract to csm package (#557)
