+++ packages/content-source-maps/.eslintrc.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+    extends: require.resolve('../../.eslintrc.cjs')
+};
