+++ packages/content-source-maps/.swrc
@@ -1,0 +1,18 @@
+{
+    "jsc": {
+        "parser": {
+        "syntax": "typescript",
+        "tsx": true
+        },
+        "target": "es2020",
+        "loose": false,
+        "minify": {
+        "compress": false,
+        "mangle": false
+        }
+    },
+    "module": {
+        "type": "es6"
+    },
+    "minify": false
+}
