+++ packages/content-source-maps/README.md
@@ -1,0 +1,3 @@
+# Content Source Maps
+
+> The base for generating Content Source Maps with Contentful
