+++ packages/content-source-maps/jest.config.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+    modulePathIgnorePatterns: ['<rootDir>/dist/'],
+    testMatch: ['**/?(*.)+(spec|test).[jt]s?(x)'],
+    transform: {
+        '^.+\\.(t|j)sx?$': '@swc/jest',
+    },
+};
