+++ packages/content-source-maps/package.json
@@ -1,0 +1,57 @@
+{
+  "name": "@contentful/content-source-maps",
+  "author": "Contentful GmbH",
+  "version": "0.1.0",
+  "description": "Base SDK for building Content Source Maps with Contentful",
+  "license": "MIT",
+  "main": "dist/index.js",
+  "type": "module",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/contentful/live-preview.git"
+  },
+  "bugs": {
+    "url": "https://github.com/contentful/live-preview/issues"
+  },
+  "homepage": "https://github.com/contentful/live-preview#readme",
+  "module": "dist/esm/index.js",
+  "types": "dist/types/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/types/index.d.ts",
+      "require": "./dist/cjs/index.js",
+      "default": "./dist/esm/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "start": "yarn concurrently \"yarn:start:*\"",
+    "start:cjs": "yarn build:cjs -w",
+    "start:esm": "yarn build:esm -w",
+    "build": "yarn concurrently \"yarn:build:*\"",
+    "build:esm": "swc src -d dist/esm",
+    "build:cjs": "swc src -d dist/cjs -C module.type=commonjs",
+    "build:types": "tsc --outDir dist/types",
+    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx",
+    "tsc": "tsc --noEmit",
+    "test": "jest --watch",
+    "test:ci": "jest",
+    "coverage": "jest --coverage",
+    "prepublish": "yarn build"
+  },
+  "dependencies": {
+    "@vercel/stega": "^0.1.0",
+    "json-pointer": "^0.6.2"
+  },
+  "devDependencies": {
+    "@swc/cli": "0.1.63",
+    "@swc/core": "1.3.100",
+    "@swc/jest": "0.2.29",
+    "@types/jest": "29.5.11",
+    "jest": "29.7.0",
+    "typescript": "5.1.6"
+  }
+}
