+++ packages/content-source-maps/src/richText.ts
