+++ packages/content-source-maps/src/utils.ts
@@ -1,0 +1,17 @@
+/**
+ * Clones the incoming element into a new one, to prevent modification on the original object
+ * Hint: It uses the structuredClone which is only available in modern browsers,
+ * for older one it uses the JSON.parse(JSON.stringify) hack.
+ */
+export function clone<T extends Record<string, unknown> | Array<unknown>>(incoming: T): T {
+  if (typeof structuredClone === 'function') {
+    return structuredClone(incoming);
+  }
+
+  try {
+    return JSON.parse(JSON.stringify(incoming));
+  } catch (err) {
+    console.warn('Failed to clone data:', incoming, err);
+    return incoming;
+  }
+}
