+++ packages/live-preview-sdk/src/inspectorMode/utils.ts
@@ -1,4 +1,4 @@
-import { decode } from '../csm/encode';
+import { decode } from '@contentful/content-source-maps';
 import { InspectorModeAttributes, InspectorModeDataAttributes } from './types';
 
 const isTaggedElement = (node?: Node | null): boolean => {
